pyproject.toml
[tool.poetry]
name = "ytdl-api"
version = "0.1.0"
description = "API for webbased youtube-dl client"
authors = ["nazarii.oleksiuk@gmail.com"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
youtube-dl = "^2020.3.6"
fastapi = "^0.52.0"
uvicorn = "^0.11.3"
pyyaml = "^5.3"
pydantic = {extras = ["dotenv"], version = "^1.4"}
python-socketio = "^4.5.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
devtools = "^0.5.1"
isort = "^4.3.21"
pytest = "^5.3.5"
flake8 = "^3.7.9"
pdbpp = "^0.10.2"
requests = "^2.23.0"
pytest-pep8 = "^1.0.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
